% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseaddress.R, R/rparseaddress.R
\name{parseaddress}
\alias{parseaddress}
\title{parseaddress()}
\usage{
parseaddress(
  address,
  check_python = TRUE,
  force_stateabb = FALSE,
  return = "char"
)

parseaddress(
  address,
  check_python = TRUE,
  force_stateabb = FALSE,
  return = "char"
)
}
\arguments{
\item{address}{string of address to parse}

\item{check_python}{TRUE or FALSE, DO NOT INSTALL MINICONDA IF PROMPTED! TRUE=check for python dependencies and install if missing, FALSE= skip check (faster if you have already have the python dependencies installed)}

\item{force_stateabb}{TRUE or FALSE, if TRUE state names are forced to abbreviation format for unified format}

\item{return}{FORMAT TYPE for RETURNED RESULTS (1. "vertical" or "v" for vertical data.frame, 2. "horizontal" or "h" for horizontal data.frame, "c" or "char" for named character vector )}
}
\description{
Get address parts from a string address.

Get address parts from a string address.
}
