% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparseaddressvec.R
\name{rparseaddressvec}
\alias{rparseaddressvec}
\title{rparseaddressvec()}
\usage{
rparseaddressvec(addressvec, return = "all")
}
\arguments{
\item{addressvec}{string vector of addresses to parse, to use a df column use df_name$address_column_name}

\item{return}{option to return single address characteristic, i.e. Zipcode}
}
\description{
Get address parts from a string addresses in a vector
}
